+++ src/Areas/Sql/Commands/SqlJsonContext.cs
@@ -4,14 +4,17 @@
 using System.Text.Json.Serialization;
 using AzureMcp.Areas.Sql.Commands.Database;
 using AzureMcp.Areas.Sql.Commands.EntraAdmin;
+using AzureMcp.Areas.Sql.Commands.FirewallRule;
 using AzureMcp.Areas.Sql.Models;
 
 namespace AzureMcp.Areas.Sql.Commands;
 
 [JsonSerializable(typeof(DatabaseShowCommand.DatabaseShowResult))]
 [JsonSerializable(typeof(EntraAdminListCommand.EntraAdminListResult))]
+[JsonSerializable(typeof(FirewallRuleListCommand.FirewallRuleListResult))]
 [JsonSerializable(typeof(SqlDatabase))]
 [JsonSerializable(typeof(SqlServerEntraAdministrator))]
+[JsonSerializable(typeof(SqlServerFirewallRule))]
 [JsonSerializable(typeof(DatabaseSku))]
 [JsonSourceGenerationOptions(
     PropertyNamingPolicy = JsonKnownNamingPolicy.CamelCase,